add cassandra protocol infer

+++ crates/l7-parser/src/cassandra/constants.rs
@@ -1,1 +1,1 @@
-pub(super) const CASSANDRA_MIN_SIZE: u32 = 9;
+pub(super) const CASSANDRA_MIN_SIZE: usize = 9;
